backend/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
sqlite3
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2
python-dotenv==1.0.0
